hw5/requirements.txt
flask
google-cloud-datastore
gunicorn
google-cloud-language==1.3.0
